Updated searchNations to return better pagination data
Note: I didn't use Relay or use cursor-based pagination since we don't
really need it. Our data source is insanely static, and we need to be
able to sort on fields like NationStrength that aren't very suitable for
cursors. Maybe it could be a thing for a future release if admin ever
gives us more real-time data updates (lol), but for now, simple
offset-based pagination should be more than fine.